fix: replace interceptor with middleware
