+++ src/app.module.ts
@@ -1,13 +1,14 @@
-import { Module } from '@nestjs/common';
+import { MiddlewareConsumer, Module, NestModule } from '@nestjs/common';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
-import { AlbumModule } from './modules/album/album.module';
 import { ArtistModule } from './modules/artist/artist.module';
-import { FavoritesModule } from './modules/favorites/favorites.module';
-import { LoggerModule } from './modules/logger/logger.module';
 import { TrackModule } from './modules/track/track.module';
+import { AlbumModule } from './modules/album/album.module';
 import { UserModule } from './modules/user/user.module';
+import { FavoritesModule } from './modules/favorites/favorites.module';
 import { PrismaModule } from './prisma/prisma.module';
+import { LoggerMiddleware } from './modules/logger/logger.middleware';
+import { LoggerModule } from './modules/logger/logger.module';
 
 @Module({
   imports: [
@@ -22,4 +23,8 @@ import { PrismaModule } from './prisma/prisma.module';
   controllers: [AppController],
   providers: [AppService],
 })
-export class AppModule {}
+export class AppModule implements NestModule {
+  configure(consumer: MiddlewareConsumer) {
+    consumer.apply(LoggerMiddleware).forRoutes('*');
+  }
+}
